requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pymongo==4.7.2
requests==2.32.3
urllib3==2.2.1
Werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
